Moved component into separate folder from app folder
