+++ learn-arai/app/classroom/page.tsx
@@ -1,4 +1,4 @@
-import Navbar from '@/components/classrooom/navbar/navbar';
+import Navbar from '@/components/module/classrooom/navbar/navbar';
 
 export default function Page() {
     return (
