+++ learn-arai/app/layout.tsx
@@ -1,8 +1,9 @@
 import type { Metadata } from 'next';
 
+import { AuthProvider } from '@/components/context/AuthContext';
+import { openSans } from '@/components/ui/fonts';
+
 import './globals.css';
-import { AuthProvider } from '@/app/context/AuthContext';
-import { openSans } from '@/app/ui/fonts';
 
 export const metadata: Metadata = {
     title: 'Create Next App',
