+++ learn-arai/app/page.tsx
@@ -2,7 +2,7 @@
 
 import { useContext } from 'react';
 
-import { AuthContext } from './context/AuthContext';
+import { AuthContext } from '@/components/context/AuthContext';
 
 export default function Home() {
     const auth = useContext(AuthContext);
